+++ client/test/payload/0/0.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+./sc-client payload --list > output.txt
+diff -Z output.txt client/test/payload/0/0.stdout